+++ src/rex.formbuilder/static/js/lib/widget/action/PickDraft.js
@@ -4,14 +4,26 @@
 
 import React from 'react';
 import autobind from 'autobind-decorator';
+import objectPath from 'object-path';
 
 import {VBox} from 'rex-widget/layout';
+import {Preloader} from 'rex-widget/ui';
+import {Fetch} from 'rex-widget/data';
 import Title from 'rex-action/lib/actions/Title';
 import {command, Types} from 'rex-action/lib/execution/Command';
 
 import {GUI, widget} from '../../index';
 
 
+@Fetch(function (props) {
+  let channels = props.channels;
+  Object.keys(props.channelFilter).forEach((key) => {
+    channels = channels.params({
+      [key]: objectPath.get(props, props.channelFilter[key])
+    });
+  });
+  return {channels};
+})
 export default class PickDraft extends React.Component {
   static commands = {
     @command(Types.Value())
@@ -36,8 +48,13 @@ export default class PickDraft extends React.Component {
      height: '100%',
      overflow: 'auto'
     };
-    let {apiBaseUrl, formPreviewerUrlTemplate, channels,
+    let {apiBaseUrl, formPreviewerUrlTemplate,
          locale, i18nBaseUrl} = this.props;
+    let {channels} = this.props.fetched;
+
+    if (channels.updating) {
+      return <Preloader />;
+    }
 
     return (
       <div
@@ -53,7 +70,7 @@ export default class PickDraft extends React.Component {
               uid={this.props.context.instrument.id}
               onDraftSelected={this.onDraftSelected}
               draftSetSelectorVerticalView={true}
-              channels={channels}
+              channels={channels.data.channels}
             />
           }
         />